Berkeley-Humanoid-Lite-Lowlevel/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "berkeley_humanoid_lite_lowlevel"
version = "2025.09.07"
description = "Low-level control for Berkeley Humanoid Lite"
authors = [
    { name = "Berkeley Humanoid Project Developers" },
    { name = "-T.K.-" },
]
maintainers = [
    { name = "Berkeley Humanoid Project Developers" },
    { name = "-T.K.-" },
]
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["robotics", "lowlevel"]
classifiers = [
    "Natural Language :: English",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "inputs>=0.5",
    "loop-rate-limiters>=1.2.0",
    "numpy>=2.2.6",
    "omegaconf>=2.3.0",
    "psutil",
    "pyserial>=3.5",
    "python-can>=4.6.1",
    "pyyaml>=6.0.3",
]

[project.urls]
repository = "https://github.com/HybridRobotics/Berkeley-Humanoid-Lite-Lowlevel"

[tool.setuptools]
packages = ["berkeley_humanoid_lite_lowlevel"]
zip-safe = false
